requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.0
h11==0.16.0
idna==3.11
mysql-connector-python==9.4.0
passlib==1.7.4
pydantic==2.12.1
pydantic_core==2.41.3
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
